LRAC-13113 Export and import hooks correctly
