+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/assets/pages/index.tsx
@@ -13,7 +13,7 @@ import {Router} from 'shared/types';
 import {sub} from 'shared/util/lang';
 import {Switch, useParams} from 'react-router-dom';
 import {useChannelContext} from 'shared/context/channel';
-import {useDataSource} from 'shared/hooks/useDataSource';
+import {useDataSource} from 'shared/hooks';
 import {User} from 'shared/util/records';
 import {withCurrentUser} from 'shared/hoc';
