+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/contacts/components/BaseListPage.tsx
@@ -19,7 +19,7 @@ import {NAME} from 'shared/util/pagination';
 import {Routes, toRoute} from 'shared/util/router';
 import {Sizes} from 'shared/util/constants';
 import {useChannelContext} from 'shared/context/channel';
-import {useDataSource} from 'shared/hooks/useDataSource';
+import {useDataSource} from 'shared/hooks';
 import {useParams} from 'react-router-dom';
 import {User} from 'shared/util/records';
