+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/individual/dashboard/pages/Distribution.tsx
@@ -14,7 +14,7 @@ import {
 import {get} from 'lodash';
 import {Routes, toRoute} from 'shared/util/router';
 import {Sizes} from 'shared/util/constants';
-import {useDataSource} from 'shared/hooks/useDataSource';
+import {useDataSource} from 'shared/hooks';
 import {useParams} from 'react-router-dom';
 import {User} from 'shared/util/records';
