+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/individual/dashboard/pages/Interests.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import StatesRenderer from 'shared/components/states-renderer/StatesRenderer';
 import URLConstants from 'shared/util/url-constants';
 import {Routes, toRoute} from 'shared/util/router';
-import {useDataSource} from 'shared/hooks/useDataSource';
+import {useDataSource} from 'shared/hooks';
 import {useParams} from 'react-router-dom';
 import {User} from 'shared/util/records';
 import {withCurrentUser} from 'shared/hoc';
