+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/shared/hooks/useQueryParams.ts
@@ -2,10 +2,8 @@ import {parse} from 'query-string';
 import {useLocation} from 'react-router-dom';
 
 // TODO: Remove this once we upgrade to react-router-dom v6
-const useQueryParams = () => {
+export const useQueryParams = () => {
 	const {search} = useLocation();
 
 	return parse(search);
 };
-
-export default useQueryParams;
