+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/sites/pages/index.tsx
@@ -16,7 +16,7 @@ import {getMatchedRoute, Routes, toRoute} from 'shared/util/router';
 import {sub} from 'shared/util/lang';
 import {Switch, useParams} from 'react-router-dom';
 import {useChannelContext} from 'shared/context/channel';
-import {useDataSource} from 'shared/hooks/useDataSource';
+import {useDataSource} from 'shared/hooks';
 import {User} from 'shared/util/records';
 import {withCurrentUser} from 'shared/hoc';
