+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/sites/touchpoints/pages/TouchpointRoutes.jsx
@@ -18,7 +18,7 @@ import {pickBy} from 'lodash';
 import {PropTypes} from 'prop-types';
 import {Switch} from 'react-router-dom';
 import {useChannelContext} from 'shared/context/channel';
-import {useDataSource} from 'shared/hooks/useDataSource';
+import {useDataSource} from 'shared/hooks';
 
 const KnownIndividuals = lazy(() =>
 	import(
